h2 porting improvements
- more shader template support
- update weapon related definitions
- weapon porting fixes
- device control porting support
- convert device control, object names, and device groups in scnr tag
- fix issue with UVs in h2v render models (uvs would be tiled 2x their normal size)

+++ TagTool/Tags/Definitions/GameObject.cs
@@ -500,7 +500,7 @@ namespace TagTool.Tags.Definitions
         }
     }
 
-    public enum GameObjectTypeHalo2 : sbyte
+    public enum GameObjectTypeHalo2 : short
     {
         Biped,
 		Vehicle,
